introduce new model folder for data models used
