+++ transformer/transformer_test.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"testing"
 
+	"github.com/ONSdigital/dp-search-api/models"
 	. "github.com/smartystreets/goconvey/convey"
 )
 
@@ -13,11 +14,11 @@ func TestLegacyTransformer(t *testing.T) {
 	Convey("Transforms unmarshalled search responses successfully", t, func() {
 		transformer := New()
 		Convey("Zero suggestions creates empty array", func() {
-			es := ESResponse{
-				Responses: []ESResponseItem{ESResponseItem{
-					Suggest: ESSuggest{
-						SearchSuggest: []ESSearchSuggest{ESSearchSuggest{
-							Options: []ESSearchSuggestOptions{},
+			es := models.ESResponse{
+				Responses: []models.ESResponseItem{models.ESResponseItem{
+					Suggest: models.ESSuggest{
+						SearchSuggest: []models.ESSearchSuggest{models.ESSearchSuggest{
+							Options: []models.ESSearchSuggestOptions{},
 						}},
 					},
 				}},
@@ -27,12 +28,12 @@ func TestLegacyTransformer(t *testing.T) {
 		})
 
 		Convey("One suggestion creates a populated array", func() {
-			es := ESResponse{
-				Responses: []ESResponseItem{ESResponseItem{
-					Suggest: ESSuggest{
-						SearchSuggest: []ESSearchSuggest{ESSearchSuggest{
-							Options: []ESSearchSuggestOptions{
-								ESSearchSuggestOptions{Text: "option1"},
+			es := models.ESResponse{
+				Responses: []models.ESResponseItem{models.ESResponseItem{
+					Suggest: models.ESSuggest{
+						SearchSuggest: []models.ESSearchSuggest{models.ESSearchSuggest{
+							Options: []models.ESSearchSuggestOptions{
+								models.ESSearchSuggestOptions{Text: "option1"},
 							},
 						}},
 					},
@@ -44,23 +45,23 @@ func TestLegacyTransformer(t *testing.T) {
 			So(sr.Suggestions[0], ShouldResemble, "option1")
 		})
 		Convey("Multiple suggestions creates a populated array incorrect order", func() {
-			es := ESResponse{
-				Responses: []ESResponseItem{ESResponseItem{
-					Suggest: ESSuggest{
-						SearchSuggest: []ESSearchSuggest{
-							ESSearchSuggest{
-								Options: []ESSearchSuggestOptions{
-									ESSearchSuggestOptions{Text: "option1"},
+			es := models.ESResponse{
+				Responses: []models.ESResponseItem{models.ESResponseItem{
+					Suggest: models.ESSuggest{
+						SearchSuggest: []models.ESSearchSuggest{
+							models.ESSearchSuggest{
+								Options: []models.ESSearchSuggestOptions{
+									models.ESSearchSuggestOptions{Text: "option1"},
 								},
 							},
-							ESSearchSuggest{
-								Options: []ESSearchSuggestOptions{
-									ESSearchSuggestOptions{Text: "option2"},
+							models.ESSearchSuggest{
+								Options: []models.ESSearchSuggestOptions{
+									models.ESSearchSuggestOptions{Text: "option2"},
 								},
 							},
-							ESSearchSuggest{
-								Options: []ESSearchSuggestOptions{
-									ESSearchSuggestOptions{Text: "option3"},
+							models.ESSearchSuggest{
+								Options: []models.ESSearchSuggestOptions{
+									models.ESSearchSuggestOptions{Text: "option3"},
 								},
 							},
 						},
@@ -127,7 +128,7 @@ func TestTransformSearchResponse(t *testing.T) {
 			actual, err := transformer.TransformSearchResponse(ctx, sampleResponse, "test-query", true)
 			So(err, ShouldBeNil)
 			So(actual, ShouldNotBeEmpty)
-			var exp, act SearchResponse
+			var exp, act models.SearchResponse
 			So(json.Unmarshal(expected, &exp), ShouldBeNil)
 			So(json.Unmarshal(actual, &act), ShouldBeNil)
 			So(act, ShouldResemble, exp)
@@ -142,7 +143,7 @@ func TestTransformSearchResponse(t *testing.T) {
 			actual, err := transformer.TransformSearchResponse(ctx, sampleResponse, "test-query", false)
 			So(err, ShouldBeNil)
 			So(actual, ShouldNotBeEmpty)
-			var exp, act SearchResponse
+			var exp, act models.SearchResponse
 			So(json.Unmarshal(expected, &exp), ShouldBeNil)
 			So(json.Unmarshal(actual, &act), ShouldBeNil)
 			So(act, ShouldResemble, exp)
@@ -157,7 +158,7 @@ func TestTransformSearchResponse(t *testing.T) {
 			actual, err := transformer.TransformSearchResponse(ctx, sampleResponse, "test query \"with quote marks\"", false)
 			So(err, ShouldBeNil)
 			So(actual, ShouldNotBeEmpty)
-			var exp, act SearchResponse
+			var exp, act models.SearchResponse
 			So(json.Unmarshal(expected, &exp), ShouldBeNil)
 			So(json.Unmarshal(actual, &act), ShouldBeNil)
 			So(act, ShouldResemble, exp)
